add generic managedBy tag, fix deletion for not managed Sn-Service associations (#356)
* add tag for Sn-Svc association, fix #317

* enforce cluster name in config, remove tag from service

+++ pkg/deploy/lattice/service_manager.go
@@ -86,6 +86,7 @@ func (m *defaultServiceManager) createAssociation(ctx context.Context, svcId *st
 	assocReq := &CreateSnSvcAssocReq{
 		ServiceIdentifier:        svcId,
 		ServiceNetworkIdentifier: aws.String(sn.ID),
+		Tags:                     m.cloud.DefaultTags(),
 	}
 	assocResp, err := m.cloud.Lattice().CreateServiceNetworkServiceAssociationWithContext(ctx, assocReq)
 	if err != nil {
@@ -187,7 +188,16 @@ func (m *defaultServiceManager) updateAssociations(ctx context.Context, svc *Ser
 	}
 
 	for _, assoc := range toDelete {
-		m.deleteAssociation(ctx, assoc.Arn)
+		isManaged, err := m.cloud.IsArnManaged(*assoc.Arn)
+		if err != nil {
+			return err
+		}
+		if isManaged {
+			err = m.deleteAssociation(ctx, assoc.Arn)
+			if err != nil {
+				return err
+			}
+		}
 	}
 
 	return nil
